requirements.txt
streamlit~=1.43.2
pandas~=2.2.3
plotly~=6.0.0
paho-mqtt~=2.1.0